feat(composer-app): Main area menu actions (#3499)
