--- packages/apps/composer-app/src/plugins/GithubMarkdownPlugin/components/MarkdownDocument/index.ts
@@ -1,5 +1,0 @@
-//
-// Copyright 2023 DXOS.org
-//
-
-export * from './MarkdownDocument';
