--- packages/apps/composer-app/src/plugins/GithubMarkdownPlugin/index.ts
@@ -1,5 +1,0 @@
-//
-// Copyright 2023 DXOS.org
-//
-
-export * from './GithubMarkdownPlugin';
