+++ packages/apps/composer-app/src/plugins/GithubPlugin/components/ExportDialog.tsx
@@ -1,0 +1,149 @@
+//
+// Copyright 2023 DXOS.org
+//
+
+import React, { useCallback, useState } from 'react';
+
+import { Button, Dialog, Trans, useTranslation } from '@dxos/aurora';
+import { ComposerModel } from '@dxos/aurora-composer';
+import { log } from '@dxos/log';
+import { Input } from '@dxos/react-appkit';
+
+import { ExportViewState, GhFileIdentifier, GhIdentifier } from '../props';
+import { useOctokitContext } from './GithubApiProviders';
+
+export const ExportDialog = ({
+  data: [_, [initialExportViewState, initialResponseUrl], model, docGhId],
+}: {
+  data: [string, [ExportViewState, string | null], ComposerModel, GhIdentifier];
+}) => {
+  const content = model.content;
+
+  const { octokit } = useOctokitContext();
+  const { t } = useTranslation('plugin-github');
+  const [exportViewState, setExportViewState] = useState<ExportViewState>(initialExportViewState);
+  const [ghResponseUrl, setGhResponseUrl] = useState<string | null>(initialResponseUrl);
+  const [ghBranchValue, setGhBranchValue] = useState('');
+  const [ghMessageValue, setGhMessageValue] = useState('');
+
+  const exportGhFileContent = useCallback(async () => {
+    if (octokit && docGhId && 'path' in docGhId && content) {
+      const branchName = ghBranchValue || t('github branch name placeholder');
+      const commitMessage = ghMessageValue || t('github commit message placeholder');
+      setExportViewState('pending');
+      try {
+        const { owner, repo, path, ref } = docGhId as GhFileIdentifier;
+        const { data: fileData } = await octokit.rest.repos.getContent({ owner, repo, path });
+        if (Array.isArray(fileData) || fileData.type !== 'file') {
+          log.error('Attempted to export to a destination in Github that is not a file.');
+          return;
+        }
+        const { sha: fileSha } = fileData;
+        const {
+          data: {
+            object: { sha: baseSha },
+          },
+        } = await octokit.rest.git.getRef({ owner, repo, ref: `heads/${ref}` });
+        await octokit.rest.git.createRef({
+          owner,
+          repo,
+          ref: `refs/heads/${branchName}`,
+          sha: baseSha,
+        });
+        await octokit.rest.repos.createOrUpdateFileContents({
+          owner,
+          repo,
+          path,
+          message: commitMessage,
+          branch: branchName,
+          sha: fileSha,
+          content: btoa(content.toString()),
+        });
+        const {
+          data: { html_url: prUrl },
+        } = await octokit.rest.pulls.create({
+          owner,
+          repo,
+          head: branchName,
+          base: ref,
+          title: commitMessage,
+        });
+        setGhResponseUrl(prUrl);
+        setExportViewState('response');
+      } catch (err) {
+        log.error('Failed to export to Github file', err);
+        setExportViewState('create-pr');
+        setGhResponseUrl(null);
+      }
+    } else {
+      log.error('Not prepared to export to Github file when requested.');
+      setGhResponseUrl(null);
+    }
+  }, [octokit, docGhId, content, ghBranchValue, ghMessageValue, t]);
+
+  return (
+    <>
+      <Dialog.Title>{t('confirm export title')}</Dialog.Title>
+      {exportViewState === 'response' ? (
+        <>
+          <p>
+            <Trans
+              {...{
+                t,
+                i18nKey: 'export to github success message',
+                values: { linkText: ghResponseUrl },
+                components: {
+                  resultStyle: <span className='text-success-600 dark:text-success-300'>_</span>,
+                  prLink: (
+                    <a
+                      href={ghResponseUrl!}
+                      target='_blank'
+                      rel='noreferrer'
+                      className='text-primary-600 dark:text-primary-300 hover:text-primary-700 dark:hover:text-primary-400'
+                    >
+                      _
+                    </a>
+                  ),
+                },
+              }}
+            />
+          </p>
+          <div role='none' className='flex justify-end'>
+            <Dialog.Close asChild>
+              <Button variant='primary' onClick={() => setExportViewState(null)}>
+                {t('done label', { ns: 'os' })}
+              </Button>
+            </Dialog.Close>
+          </div>
+        </>
+      ) : (
+        <div role='none' className='mbs-2 space-b-2'>
+          <Input
+            disabled={exportViewState === 'pending'}
+            label={t('github branch name label')}
+            placeholder={t('github branch name placeholder')}
+            value={ghBranchValue}
+            onChange={({ target: { value } }) => setGhBranchValue(value)}
+            slots={{ input: { autoFocus: true, className: 'font-mono' } }}
+          />
+          <Input
+            disabled={exportViewState === 'pending'}
+            label={t('github commit message label')}
+            placeholder={t('github commit message placeholder')}
+            value={ghMessageValue}
+            onChange={({ target: { value } }) => setGhMessageValue(value)}
+            size='textarea'
+          />
+          <div role='none' className='flex justify-end gap-2'>
+            <Dialog.Close asChild>
+              <Button onClick={() => setExportViewState(null)}>{t('close label', { ns: 'os' })}</Button>
+            </Dialog.Close>
+            <Button disabled={exportViewState === 'pending'} variant='primary' onClick={() => exportGhFileContent()}>
+              {t('continue label', { ns: 'os' })}
+            </Button>
+          </div>
+        </div>
+      )}
+    </>
+  );
+};
