+++ packages/apps/composer-app/src/plugins/GithubPlugin/components/GithubApiProviders/PatInput.tsx
@@ -7,7 +7,6 @@ import React, { PropsWithChildren, useEffect, useState } from 'react';
 
 import { Link, Input, Trans, useTranslation, useId } from '@dxos/aurora';
 import { getSize, mx } from '@dxos/aurora-theme';
-import { useSplitViewContext } from '@dxos/react-surface';
 
 import { useOctokitContext } from './OctokitProvider';
 
@@ -28,20 +27,17 @@ const ExternalLink = ({ children }: PropsWithChildren<{}>) => {
 };
 
 export const PatInput = () => {
-  const { t } = useTranslation('composer');
+  const { t } = useTranslation('plugin-github');
   const { pat, setPat } = useOctokitContext();
   const [patValue, setPatValue] = useState(pat);
-  const { dialogOpen } = useSplitViewContext();
 
   useEffect(() => {
     setPatValue(pat);
   }, [pat]);
 
   useEffect(() => {
-    if (!dialogOpen) {
-      void setPat(patValue);
-    }
-  }, [dialogOpen]);
+    void setPat(patValue);
+  }, [patValue]);
 
   return (
     <Input.Root>
