+++ packages/apps/composer-app/src/plugins/GithubPlugin/components/index.ts
@@ -4,5 +4,8 @@
 
 export * from './GithubApiProviders';
 export * from './GithubEchoResolverProviders';
-export * from './StandaloneMain';
+export * from './MarkdownActions';
+export * from './ExportDialog';
+export * from './ImportDialog';
+export * from './UrlDialog';
 export * from './EmbeddedMain';
