+++ packages/apps/composer-app/src/plugins/GithubPlugin/hooks/index.ts
@@ -2,4 +2,4 @@
 // Copyright 2023 DXOS.org
 //
 
-export * from './StandaloneMain';
+export * from './useDocGhId';
