+++ packages/apps/composer-app/src/plugins/MarkdownPlugin/components/StandaloneLayout.tsx
@@ -2,33 +2,28 @@
 // Copyright 2023 DXOS.org
 //
 
-import { DotsThreeVertical, FilePlus } from '@phosphor-icons/react';
-import React, { Dispatch, HTMLAttributes, PropsWithChildren, ReactNode, SetStateAction } from 'react';
-import { FileUploader } from 'react-drag-drop-files';
+import { DotsThreeVertical } from '@phosphor-icons/react';
+import React, { HTMLAttributes, PropsWithChildren, RefObject } from 'react';
 
 import { Button, DropdownMenu, ThemeContext, Main, useThemeContext, useTranslation } from '@dxos/aurora';
+import { ComposerModel, MarkdownComposerRef } from '@dxos/aurora-composer';
 import { defaultBlockSeparator, getSize, mx, osTx } from '@dxos/aurora-theme';
-import { Dialog, Input } from '@dxos/react-appkit';
+import { Input } from '@dxos/react-appkit';
 import { observer } from '@dxos/react-client';
+import { Surface } from '@dxos/react-surface';
 
 import { MarkdownProperties } from './MarkdownMain';
 
 export const StandaloneLayout = observer(
   ({
     children,
-    id,
+    model,
     properties,
-    dropdownMenuContent,
-    handleFileImport,
-    fileImportDialogOpen,
-    setFileImportDialogOpen,
+    editorRef,
   }: PropsWithChildren<{
-    id: string;
+    model: ComposerModel;
     properties: MarkdownProperties;
-    dropdownMenuContent?: ReactNode;
-    handleFileImport?: (file: File) => Promise<void>;
-    fileImportDialogOpen?: boolean;
-    setFileImportDialogOpen?: Dispatch<SetStateAction<boolean>>;
+    editorRef: RefObject<MarkdownComposerRef>;
   }>) => {
     const { t } = useTranslation('composer');
     const themeContext = useThemeContext();
@@ -37,7 +32,7 @@ export const StandaloneLayout = observer(
         <div role='none' className='mli-auto max-is-[60rem] min-bs-[100vh] bg-white dark:bg-neutral-925 flex flex-col'>
           <div role='none' className='flex items-center gap-2 pis-0 pointer-fine:pis-8 lg:pis-0 pointer-fine:lg:pis-0'>
             <Input
-              key={id}
+              key={model.id}
               variant='subdued'
               label={t('document title label')}
               labelVisuallyHidden
@@ -61,7 +56,7 @@ export const StandaloneLayout = observer(
                 </DropdownMenu.Trigger>
                 <DropdownMenu.Portal>
                   <DropdownMenu.Content sideOffset={10} classNames='z-10'>
-                    {dropdownMenuContent}
+                    <Surface data={[model, properties, editorRef]} role='menuitem' />
                     <DropdownMenu.Arrow />
                   </DropdownMenu.Content>
                 </DropdownMenu.Portal>
@@ -71,30 +66,30 @@ export const StandaloneLayout = observer(
           <div role='separator' className={mx(defaultBlockSeparator, 'mli-3 opacity-50')} />
           {children}
         </div>
-        <ThemeContext.Provider value={{ ...themeContext, tx: osTx }}>
-          {handleFileImport && (
-            <Dialog
-              open={fileImportDialogOpen}
-              onOpenChange={setFileImportDialogOpen}
-              title={t('confirm import title')}
-              slots={{ overlay: { classNames: 'backdrop-blur-sm' } }}
-            >
-              <p className='mlb-4'>{t('confirm import body')}</p>
-              <FileUploader
-                types={['md']}
-                classes='block mlb-4 p-8 border-2 border-dashed border-neutral-500/50 rounded flex items-center justify-center gap-2 cursor-pointer'
-                dropMessageStyle={{ border: 'none', backgroundColor: '#EEE' }}
-                handleChange={handleFileImport}
-              >
-                <FilePlus weight='duotone' className={getSize(8)} />
-                <span>{t('upload file message')}</span>
-              </FileUploader>
-              <Button classNames='block is-full' onClick={() => setFileImportDialogOpen?.(false)}>
-                {t('cancel label', { ns: 'appkit' })}
-              </Button>
-            </Dialog>
-          )}
-        </ThemeContext.Provider>
+        {/* <ThemeContext.Provider value={{ ...themeContext, tx: osTx }}> */}
+        {/*  {handleFileImport && ( */}
+        {/*    <Dialog */}
+        {/*      open={fileImportDialogOpen} */}
+        {/*      onOpenChange={setFileImportDialogOpen} */}
+        {/*      title={t('confirm import title')} */}
+        {/*      slots={{ overlay: { classNames: 'backdrop-blur-sm' } }} */}
+        {/*    > */}
+        {/*      <p className='mlb-4'>{t('confirm import body')}</p> */}
+        {/*      <FileUploader */}
+        {/*        types={['md']} */}
+        {/*        classes='block mlb-4 p-8 border-2 border-dashed border-neutral-500/50 rounded flex items-center justify-center gap-2 cursor-pointer' */}
+        {/*        dropMessageStyle={{ border: 'none', backgroundColor: '#EEE' }} */}
+        {/*        handleChange={handleFileImport} */}
+        {/*      > */}
+        {/*        <FilePlus weight='duotone' className={getSize(8)} /> */}
+        {/*        <span>{t('upload file message')}</span> */}
+        {/*      </FileUploader> */}
+        {/*      <Button classNames='block is-full' onClick={() => setFileImportDialogOpen?.(false)}> */}
+        {/*        {t('cancel label', { ns: 'appkit' })} */}
+        {/*      </Button> */}
+        {/*    </Dialog> */}
+        {/*  )} */}
+        {/* </ThemeContext.Provider> */}
       </Main.Content>
     );
   },
